added drag controls
