+++ src/App.js
@@ -1,152 +1,26 @@
 import "./App.css";
 //useThree is used to allow OrbitControls to use the react-three-fiber elements
 //extend is used to allow OrbitControls to be used in JSX
-import {
-  Canvas,
-  useFrame,
-  extend,
-  useThree,
-  useLoader,
-} from "@react-three/fiber";
-import { OrbitControls } from "three/examples/jsm/controls/OrbitControls";
+import { Canvas, useFrame, useThree, useLoader } from "@react-three/fiber";
+
 import * as THREE from "three";
 
 //Hook to give access to an element
-import { useRef, Suspense } from "react";
+import { Suspense } from "react";
 import { useEffect } from "react";
 
-extend({ OrbitControls });
-
-const Orbit = () => {
-  const { camera, gl } = useThree();
-  useEffect(() => {
-    const controls = new OrbitControls(camera, gl.domElement);
-    return () => {
-      controls.dispose();
-    };
-  }, [camera, gl]);
-  return null;
-
-  //camera and gl props are passed into OrbitControls
-  // return <OrbitControls args={[camera, gl.domElement]} />;
-};
-
-const Box = (props) => {
-  //Using useRef to gain reference to the react-three-fiber element (3D Object)
-  const ref = useRef();
-
-  // useLoader used to load the texture
-  const texture = useLoader(THREE.TextureLoader, "/wood.jpg");
-
-  //useFrame takes a call back that runs on every render
-  useFrame((state) => {
-    //current gives access to the referenced object
-    ref.current.rotation.x += 0.01;
-    ref.current.rotation.y += 0.01;
-  });
-
-  //function
-  const handlePointerDown = (e) => {
-    e.object.active = true;
-    if (window.activeMesh) {
-      scaleDown(window.activeMesh);
-      window.activeMesh.active = false;
-    }
-    window.activeMesh = e.object;
-  };
-
-  const handlePointerEnter = (e) => {
-    e.object.scale.x = 1.5;
-    e.object.scale.y = 1.5;
-    e.object.scale.z = 1.5;
-  };
-
-  const handlePointerLeave = (e) => {
-    if (!e.object.active) {
-      scaleDown(e.object);
-    }
-  };
-
-  const scaleDown = (object) => {
-    object.scale.x = 1;
-    object.scale.y = 1;
-    object.scale.z = 1;
-  };
-
-  return (
-    <mesh
-      ref={ref}
-      {...props}
-      castShadow
-      onPointerDown={handlePointerDown}
-      onPointerEnter={handlePointerEnter}
-      onPointerLeave={handlePointerLeave}
-    >
-      <boxBufferGeometry args={[1, 1, 1]} />
-      <meshPhysicalMaterial map={texture} />
-    </mesh>
-  );
-};
-const Background = (props) => {
-  const texture = useLoader(THREE.TextureLoader, "/autoshop.jpg");
-
-  const { gl } = useThree();
-
-  const formatted = new THREE.WebGLCubeRenderTarget(
-    texture.image.height
-  ).fromEquirectangularTexture(gl, texture);
-
-  return <primitive attach="background" object={formatted.texture} />;
-};
-
-const Floor = (props) => {
-  return (
-    <mesh {...props} receiveShadow>
-      <boxBufferGeometry args={[20, 1, 10]} />
-      <meshPhysicalMaterial />
-    </mesh>
-  );
-};
-
-const Bulb = (props) => {
-  return (
-    <mesh {...props}>
-      <pointLight castShadow />
-      <sphereBufferGeometry args={[0.2, 20, 20]} />
-
-      <meshPhongMaterial emissive="yellow" />
-    </mesh>
-  );
-};
+import Orbit from "./components/Orbit";
+import Box from "./components/Box";
+import Background from "./components/Background";
+import Floor from "./components/Floor";
+import Bulb from "./components/Bulb";
+import ColorPicker from "./components/ColorPicker";
+import Dragable from "./components/Dragable";
 
 function App() {
-  const handleClick = (e) => {
-    if (!window.activeMesh) return;
-
-    window.activeMesh.material.color = new THREE.Color(
-      e.target.style.background
-    );
-  };
-
   return (
     <div style={{ height: "100vh", width: "100vw" }}>
-      {/* Color Selector */}
-      <div style={{ position: "absolute", zIndex: 1 }}>
-        <div
-          onClick={handleClick}
-          style={{ background: "blue", height: 50, width: 50 }}
-        ></div>
-        <div
-          onClick={handleClick}
-          style={{ background: "yellow", height: 50, width: 50 }}
-        ></div>
-        <div
-          onClick={handleClick}
-          style={{ background: "white", height: 50, width: 50 }}
-        ></div>
-      </div>
-
-      {/* End of Color Selector */}
+      <ColorPicker />
       <Canvas
         shadows
         style={{ background: "black" }}
@@ -155,14 +29,16 @@ function App() {
         {/* <fog attach="fog" args={["white", 1, 10]} /> */}
         <ambientLight intensity={0.2} />
         <Bulb position={[0, 3, 0]} />
-        <Orbit />
+        <Orbit attach="orbitControls" />
         <axesHelper args={[5]} />
-        <Suspense fallback={null}>
-          <Box position={[-4, 1, 0]} />
-        </Suspense>
-        <Suspense fallback={null}>
-          <Box position={[0, 1, 0]} />
-        </Suspense>
+        <Dragable>
+          <Suspense fallback={null}>
+            <Box position={[-4, 1, 0]} />
+          </Suspense>
+          <Suspense fallback={null}>
+            <Box position={[0, 1, 0]} />
+          </Suspense>
+        </Dragable>
         <Suspense fallback={null}>
           <Background />
         </Suspense>
