+++ src/components/Box.js
@@ -1,0 +1,62 @@
+import { useRef } from "react";
+import { useFrame, useLoader } from "@react-three/fiber";
+import * as THREE from "three";
+
+const Box = (props) => {
+  //Using useRef to gain reference to the react-three-fiber element (3D Object)
+  const ref = useRef();
+
+  // useLoader used to load the texture
+  const texture = useLoader(THREE.TextureLoader, "/wood.jpg");
+
+  //useFrame takes a call back that runs on every render
+  useFrame((state) => {
+    //current gives access to the referenced object
+    ref.current.rotation.x += 0.01;
+    ref.current.rotation.y += 0.01;
+  });
+
+  //function
+  const handlePointerDown = (e) => {
+    e.object.active = true;
+    if (window.activeMesh) {
+      scaleDown(window.activeMesh);
+      window.activeMesh.active = false;
+    }
+    window.activeMesh = e.object;
+  };
+
+  const handlePointerEnter = (e) => {
+    e.object.scale.x = 1.5;
+    e.object.scale.y = 1.5;
+    e.object.scale.z = 1.5;
+  };
+
+  const handlePointerLeave = (e) => {
+    if (!e.object.active) {
+      scaleDown(e.object);
+    }
+  };
+
+  const scaleDown = (object) => {
+    object.scale.x = 1;
+    object.scale.y = 1;
+    object.scale.z = 1;
+  };
+
+  return (
+    <mesh
+      ref={ref}
+      {...props}
+      castShadow
+      onPointerDown={handlePointerDown}
+      onPointerEnter={handlePointerEnter}
+      onPointerLeave={handlePointerLeave}
+    >
+      <boxBufferGeometry args={[1, 1, 1]} />
+      <meshPhysicalMaterial map={texture} />
+    </mesh>
+  );
+};
+
+export default Box;
